// swiyu-trust-management-ui/src/app/api/generated/model/task-list-item.ts
/**
 * SWIYU Trust Management SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {TaskAction} from './task-action';
import {TaskStatus} from './task-status';

export interface TaskListItem {
  id: string;
  partnerName: {[key: string]: string};
  submittedAt: string;
  dueAt: string;
  state: TaskStatus;
  assignee?: string;
  taskType: string;
  allowedActions: Set<TaskAction>;
}
export namespace TaskListItem {}