// swiyu-trust-management-ui/src/app/api/generated/model/task-status.ts
/**
 * SWIYU Trust Management SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export const TaskStatus = {
  Rejected: 'REJECTED',
  Accepted: 'ACCEPTED',
  Opened: 'OPENED',
  InformationRequested: 'INFORMATION_REQUESTED',
  Resubmitted: 'RESUBMITTED'
} as const;
export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];