// swiyu-trust-management-ui/src/app/api/generated/model/trust-onboarding-reject-reason.ts
/**
 * SWIYU Trust Management SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export const TrustOnboardingRejectReason = {
  IncompleteInformation: 'INCOMPLETE_INFORMATION',
  InaccurateInformation: 'INACCURATE_INFORMATION',
  OutdatedInformation: 'OUTDATED_INFORMATION',
  IdentityVerificationFailure: 'IDENTITY_VERIFICATION_FAILURE',
  LackOfAuthorization: 'LACK_OF_AUTHORIZATION',
  TechnicalIssues: 'TECHNICAL_ISSUES',
  DuplicateApplication: 'DUPLICATE_APPLICATION',
  NoResponseFromApplicant: 'NO_RESPONSE_FROM_APPLICANT',
  FraudulentActivity: 'FRAUDULENT_ACTIVITY',
  Other: 'OTHER'
} as const;
export type TrustOnboardingRejectReason =
  (typeof TrustOnboardingRejectReason)[keyof typeof TrustOnboardingRejectReason];